Multiple levels of specialization are now possible.
One problem that arised is the type of the 'next' member of the following snippet, after the type is specialized for <value: int>:

type Element
{
	var Element next;
	var value;
}

Should 'next' now be of type Element<value: int> or simply Element? Should there be a keyword to explicitly ask for a member of the same type as the enclosing type?