code gen for custom operators works
Fixed an issue where the names of custom operators would not be generated in a C-compatible way. Now symbols in the names are converted to their equivalent hexadecimal representation. Also added automatic generation of the int main() stub that calls the actual main function generated. This will turn out to be very handy when converting argc and argv to a native array before passing it to the main function.