first steps of rewriting CodeGenerator
Added the CodeGen2 class and some helper classes to surround it. Eventually it
will replace the current CodeGenerator, being built from the ground up with
dependencies and sqlite3 in mind. The new scheme is much more convenient, with
the generator code for individual nodes sitting in separate files to provide a
better overview of the system.