fixed ast-gen interface order
The order of interfaces ast-gen generates for AST nodes is now deterministic.
The interfaces were stored in a set of pointers, which could cause them to be
ordered in any arbitrary way, since pointers returned from operator new need
not be monotonically increasing. The set that keeps the pointers now uses a
custom comparison operator that compares the interfaces based on their names.