fixed linking issue in driver
The driver classes are compiled into an intermediate library, which was not
linked against the library containing the AST code. On Mac OS X this was no
issue since all executables that used libdriver explicitly linked against
libast as well. On Linux, however, this is an issue since the order in which
linking is performed matters. The libdriver target now properly links against
the AST on its own and should therefore eliminate these issues.