included tests/ in ctest setup
The `tests/` subdirectory now has its own CMakeLists.txt that configures a
parse test for each of the `*.mw` files in there. This allows robust regression
testing on the parser and compiler in general. If a feature gets added to the
compiler, a sample source file that breaks if the feature is not present should
be placed in `tests/`. This prevent later changes from breaking existing code.