added UnionType codegen
First steps towards issue #12. The code for union types is now generated, but
cannot be used by anything. The AssignmentExpr needs to be extended to select
the appropriate active union member when assigning to a variable of union type.
Or rather have a cast that casts e.g. `Int` to `Int|Real` by appropriately
selecting the active union member.