added test case for UnionType
The file test6.mw contains a very simple piece of code that has a variable which
is of UnionType. The constant 0 is assigned to the variable, causing the
frontend to insert a UnionMappedType that wraps the Int in a Int|Real. The
backend is required to accomodate all these features in C code.