added regression test to test6.mw
Line 5 of tests/test6.mw reveals an issue in IfCaseExpr where the individual
branch expressions would not be assigned the final requiredType. E.g. in an
IfCaseExpr that returns Int in one and Real in the other branch, the required
type of both branches remains "*", instead of being set to "Int|Real" as it
ought to be.