added codegen for UnionMappedType
Added infrastructure that allows type mappings to generate additional code for
an expression. At the moment only UnionMappedType is implemented, which wraps
the code of an expression inside an inline struct initializer. Other type
mappings that are likely to come up in the future include casts and interface
mappings.

There is an unrelated issue in IfCaseExpr though, as detailed in the previous
commit.