added versioning
The index file generated by the FileSourceRepository now contains version
information at the beginning of the file. This should allow the tool to
open older versions of the file.

Also added a version stamp for the entire project. The file config.hpp.in
is configured by CMake, replacing all known variables. As such, it defines
the VERSION_{MAJOR,MINOR,PATCH} integers, and the VERSION_STRING. Pretty
handy.

mwc-src now features a -h and --help option to display the help page. Also
available is the --version option that prints the current VERSION_STRING.