native: funcs now pull in includes
Native functions declared in the code may now declare the headers that are
required to call them. The backend keeps track of include files required for
each fragment and generates include statements accordingly. The grammar for
native functions is also a bit more cleaned up now, allowing for more convenient
declaration of multiple functions. This might be a good reference for future
grammar changes.