compiler: added native type handling
The compiler now allows for native C types to be used in Maxwell code. It will
generate the appropriate output, allowing for proper calls to standard C library
functions such as malloc or puts/gets. Casts are still missing however, making
it almost impossible to use these things properly. As soon as casts are in
place, writing a basic standard library for strings and arrays should be a piece
of cake.