impl cast expr
Implemented the CastExpr, which allows the programmer to change the assumed type
of an expression. This is very handy for the low-level hackery required in the
beginning. It is now possible to implement a rudimentary dynamic string type.
Combined with objects/values/structs, this should allow for the first "classes"
to be implemented.