added location tracking to lexer and parser
Modified the location tracking code of the lexer and parser to now use the types
defined in maxwell/location.hpp. In conjunction with a source file repository,
these types shall make it possible to produce the exact position and wording of
the location in the source file they refer to. This shall allow for proper error
messages to be produced via the diagnostics subsystem yet to be written.