Add extensions to <iterator>
Create a header maxwell/iterator.hpp that includes <iterator> and
defines extensions to the standard iterators. Define a mapping_iterator
class that wraps around any iterator and applies a predefined mapping
upon being dereferenced. Define a dereference mapping that simply
dereferences the argument passed to it. Define a dereferencing_iterator
that combines the previous two and allows iteration over a container of
std::unique_ptr where the iterator resolves to the bare pointer.

* Define mapping_iterator
* Define dereference mapping
* Define dereferencing_iteratpr