Add range class
Define a maxwell::range class that represents a slice or range of
elements within a container as a pair of iterators. This simplifies the
passing around of iterators, especially as a return value of a function.
Instead of returning a const ref to the container, member functions may
return a range covering said container's elements, making code that uses
said function much more container type agnostic.

* Add maxwell/range.hpp