Add source range tracking to AST nodes
Add two SourceRange members to the ast::Node base class, that allows the
node to keep track of its location within the source file. Having two
separate ranges allows the node to track the actual range it covered,
and the (potentially much smaller) range that shall be used to refer to
it from diagnostic messages.

Also add the corresponding encoding and decoding facilities.

* Add range, getRange(), setRange() to ast::Node
* Add referenceRange, getReferenceRange(), setReferenceRange() to
  ast::Node
* Adjust ast/Coder.hpp, ast/Serializer.cpp, gen/ast-gen.cpp accordingly